formattedTimeのテストを描いた

+++ MECap/MECap/Extensions/Double.swift
@@ -10,13 +10,12 @@ extension Double {
         let hours = Int(self) / 3600
         let minutes = (Int(self) % 3600) / 60
         let seconds = self.truncatingRemainder(dividingBy: 60)
-        let underFloat = Int(seconds.truncatingRemainder(dividingBy: 1)*10)
-        
+        let underFloat = seconds.truncatingRemainder(dividingBy: 1)*10
         
         if hours == 0 {
-            return String(format: "%02d:%02d.%01d", minutes, Int(seconds), underFloat)
+            return String(format: "%02d:%02d.%1d", minutes, Int(seconds), Int(underFloat))
         } else {
-            return String(format: "%02d:%02d:%02d.%01d", hours, minutes, seconds, underFloat)
+            return String(format: "%02d:%02d:%02d.%1d", hours, minutes, Int(seconds), Int(underFloat))
         }
     }
 }
